filled trip-controller #1
